added formatting to freq measurement in MHz

+++ util/freq-utils-m.h
@@ -47,5 +47,5 @@ typedef struct cpu_freq_data{
 void init_unit_data(unit_data *data);
 CFDictionaryRef sample(unit_data *data, int time);
 uint64_t *get_state_res(CFDictionaryRef cpu_delta, int core_id);
-uint64_t get_frequency(CFDictionaryRef cpu_delta, int core_id);
+float get_frequency(CFDictionaryRef cpu_delta, int core_id);
 #endif
